backend/requirements.txt
Django==4.2.21
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
